+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 0.0.3 WIP
 
+* Adds player's game registration and game listing endpoint from playercenter-backend
 * Refactors galaxy-spiral to spiral-galaxy
 
 # 0.0.2